feat: Pass through headers and query parameters
Now headers and query parameters are passed through
so, for example Bell can work that send the `location`
header and at second stage sends a query parameter with
a `code`.

Some settings has also been added, such as if the headers
should be filtered (content-encoding and transfer-encoding)
as well as an option if the body should be stringified or not.
Both options default to true: `{filterHeaders: true, stringifyBody: true}`.

Added a new example showing of usage of Bell.